+++ spec/rails_helper.rb
@@ -54,6 +54,7 @@ RSpec.configure do |config|
   # config.filter_gems_from_backtrace("gem name")
 
   config.include ActiveJob::TestHelper
+  config.include ActiveSupport::Testing::TimeHelpers
 end
 
 OmniAuth.config.test_mode = true